+++ src/access/index.ts
@@ -1,3 +1,4 @@
 export * from './isAdmin';
 export * from './isAdminOrSelf';
 export * from './isAdminSelfOrParty';
+export * from './isAuthed';