+++ src/blocks/ButtonLink.ts
@@ -1,48 +1,16 @@
 import { Block } from 'payload/types';
 
 import { link } from '../fields';
+import { color } from '../fields/color';
 
 const ButtonLink: Block = {
   slug: 'buttonLink',
   fields: [
     {
       name: 'icon',
-      label: 'Icon',
       type: 'text',
     },
-    {
-      name: 'color',
-      label: 'Color',
-      type: 'select',
-      required: true,
-      defaultValue: 'neutral',
-      options: [
-        {
-          label: 'Neutral',
-          value: 'neutral',
-        },
-        {
-          label: 'Neutral Variant',
-          value: 'neutral-variant',
-        },
-        {
-          label: 'Primary',
-          value: 'primary',
-        },
-        {
-          label: 'Secondary',
-          value: 'secondary',
-        },
-        {
-          label: 'Tertiary',
-          value: 'tertiary',
-        },
-        {
-          label: 'Danger',
-          value: 'danger',
-        },
-      ],
-    },
+    color,
     link,
   ],
 };