+++ src/fields/link.ts
@@ -1,7 +1,8 @@
 import { Field } from 'payload/types';
 
 export const link: Field = {
-  type: 'row',
+  name: 'link',
+  type: 'group',
   fields: [
     {
       name: 'type',
@@ -46,7 +47,7 @@ export const link: Field = {
       name: 'reference',
       label: 'Page',
       type: 'relationship',
-      relationTo: 'pages',
+      relationTo: ['pages', 'protected-pages'],
       required: true,
       maxDepth: 1,
       admin: {
@@ -61,6 +62,10 @@ export const link: Field = {
       admin: {
         condition: (_, siblingData) => siblingData?.type === 'external',
       },
+    },
+    {
+      name: 'icon',
+      type: 'text',
     },
   ],
 };