+++ src/payload-types.ts
@@ -11,6 +11,7 @@ export interface Config {
     guests: Guest;
     pages: Page;
     parties: Party;
+    'protected-pages': ProtectedPage;
     relations: Relation;
     sides: Side;
     users: User;
@@ -80,6 +81,34 @@ export interface Page {
   content: {
     layout?: (
       | {
+          title: string;
+          icon: string;
+          content: {
+            [k: string]: unknown;
+          }[];
+          color: 'neutral' | 'neutral-variant' | 'primary' | 'secondary' | 'tertiary' | 'danger';
+          link: {
+            type: 'reference' | 'external';
+            newTab?: boolean;
+            text: string;
+            reference:
+              | {
+                  value: string | Page;
+                  relationTo: 'pages';
+                }
+              | {
+                  value: string | ProtectedPage;
+                  relationTo: 'protected-pages';
+                };
+            url: string;
+            icon?: string;
+          };
+          width?: 'full' | 'max';
+          id?: string;
+          blockName?: string;
+          blockType: 'alert';
+        }
+      | {
           width?: 'full' | 'max';
           content: {
             [k: string]: unknown;
@@ -105,6 +134,57 @@ export interface Page {
           border: 'none' | 'left' | 'right';
           layout?: (
             | {
+                title: string;
+                icon: string;
+                content: {
+                  [k: string]: unknown;
+                }[];
+                color: 'neutral' | 'neutral-variant' | 'primary' | 'secondary' | 'tertiary' | 'danger';
+                link: {
+                  type: 'reference' | 'external';
+                  newTab?: boolean;
+                  text: string;
+                  reference:
+                    | {
+                        value: string | Page;
+                        relationTo: 'pages';
+                      }
+                    | {
+                        value: string | ProtectedPage;
+                        relationTo: 'protected-pages';
+                      };
+                  url: string;
+                  icon?: string;
+                };
+                width?: 'full' | 'max';
+                id?: string;
+                blockName?: string;
+                blockType: 'alert';
+              }
+            | {
+                icon?: string;
+                color: 'neutral' | 'neutral-variant' | 'primary' | 'secondary' | 'tertiary' | 'danger';
+                link: {
+                  type: 'reference' | 'external';
+                  newTab?: boolean;
+                  text: string;
+                  reference:
+                    | {
+                        value: string | Page;
+                        relationTo: 'pages';
+                      }
+                    | {
+                        value: string | ProtectedPage;
+                        relationTo: 'protected-pages';
+                      };
+                  url: string;
+                  icon?: string;
+                };
+                id?: string;
+                blockName?: string;
+                blockType: 'buttonLink';
+              }
+            | {
                 width?: 'full' | 'max';
                 content: {
                   [k: string]: unknown;
@@ -113,17 +193,138 @@ export interface Page {
                 blockName?: string;
                 blockType: 'content';
               }
+          )[];
+          blockName?: string;
+          blockType: 'section';
+        }
+    )[];
+  };
+  updatedAt: string;
+  createdAt: string;
+  _status?: 'draft' | 'published';
+}
+export interface ProtectedPage {
+  id: string;
+  name: string;
+  slug?: string;
+  meta: {
+    title: string;
+    description: string;
+  };
+  content: {
+    layout?: (
+      | {
+          title: string;
+          icon: string;
+          content: {
+            [k: string]: unknown;
+          }[];
+          color: 'neutral' | 'neutral-variant' | 'primary' | 'secondary' | 'tertiary' | 'danger';
+          link: {
+            type: 'reference' | 'external';
+            newTab?: boolean;
+            text: string;
+            reference:
+              | {
+                  value: string | Page;
+                  relationTo: 'pages';
+                }
+              | {
+                  value: string | ProtectedPage;
+                  relationTo: 'protected-pages';
+                };
+            url: string;
+            icon?: string;
+          };
+          width?: 'full' | 'max';
+          id?: string;
+          blockName?: string;
+          blockType: 'alert';
+        }
+      | {
+          width?: 'full' | 'max';
+          content: {
+            [k: string]: unknown;
+          }[];
+          id?: string;
+          blockName?: string;
+          blockType: 'content';
+        }
+      | {
+          titleOne: string;
+          titleTwo: string;
+          subtitle: string;
+          id?: string;
+          blockName?: string;
+          blockType: 'hero';
+        }
+      | {
+          id: string;
+          title: string;
+          description?: {
+            [k: string]: unknown;
+          }[];
+          border: 'none' | 'left' | 'right';
+          layout?: (
+            | {
+                title: string;
+                icon: string;
+                content: {
+                  [k: string]: unknown;
+                }[];
+                color: 'neutral' | 'neutral-variant' | 'primary' | 'secondary' | 'tertiary' | 'danger';
+                link: {
+                  type: 'reference' | 'external';
+                  newTab?: boolean;
+                  text: string;
+                  reference:
+                    | {
+                        value: string | Page;
+                        relationTo: 'pages';
+                      }
+                    | {
+                        value: string | ProtectedPage;
+                        relationTo: 'protected-pages';
+                      };
+                  url: string;
+                  icon?: string;
+                };
+                width?: 'full' | 'max';
+                id?: string;
+                blockName?: string;
+                blockType: 'alert';
+              }
             | {
                 icon?: string;
                 color: 'neutral' | 'neutral-variant' | 'primary' | 'secondary' | 'tertiary' | 'danger';
-                type: 'reference' | 'external';
-                newTab?: boolean;
-                text: string;
-                reference: string | Page;
-                url: string;
+                link: {
+                  type: 'reference' | 'external';
+                  newTab?: boolean;
+                  text: string;
+                  reference:
+                    | {
+                        value: string | Page;
+                        relationTo: 'pages';
+                      }
+                    | {
+                        value: string | ProtectedPage;
+                        relationTo: 'protected-pages';
+                      };
+                  url: string;
+                  icon?: string;
+                };
                 id?: string;
                 blockName?: string;
                 blockType: 'buttonLink';
+              }
+            | {
+                width?: 'full' | 'max';
+                content: {
+                  [k: string]: unknown;
+                }[];
+                id?: string;
+                blockName?: string;
+                blockType: 'content';
               }
           )[];
           blockName?: string;
@@ -152,11 +353,22 @@ export interface User {
 export interface NavMenu {
   id: string;
   links?: {
-    type: 'reference' | 'external';
-    newTab?: boolean;
-    text: string;
-    reference: string | Page;
-    url: string;
+    link: {
+      type: 'reference' | 'external';
+      newTab?: boolean;
+      text: string;
+      reference:
+        | {
+            value: string | Page;
+            relationTo: 'pages';
+          }
+        | {
+            value: string | ProtectedPage;
+            relationTo: 'protected-pages';
+          };
+      url: string;
+      icon?: string;
+    };
     id?: string;
   }[];
   updatedAt?: string;